Renamed control folder to controller folder
